+++ frontend/src/components/client/layout/NavbarClient.tsx
@@ -12,8 +12,8 @@ import { GetProductsParams } from "@/services/productService";
 import { Category } from "@/types";
 import { Transition } from "@headlessui/react";
 import { useTranslations } from "next-intl";
-import Link from "next/link";
-import { usePathname, useRouter, useSearchParams } from "next/navigation";
+import { useSearchParams } from "next/navigation";
+import { Link, usePathname, useRouter } from "@/i18n/navigation";
 import {
   Fragment,
   useCallback,
